refactored microcode generation
Refactored microcode generation so as to be driven by a YAML configuration file. This way, as the breadboard computer evolves, we don't have to have multiple versions of the generation python. The differences between breadboard computer versions are captured in a configuration file.
